+++ python/sglang/srt/utils.py
@@ -50,6 +50,7 @@ from fastapi.responses import ORJSONResponse
 from packaging import version as pkg_version
 from starlette.routing import Mount
 from torch import nn
+from torch.distributed.tensor import DTensor, Shard, Replicate
 from torch.func import functional_call
 from torch.library import Library
 from torch.profiler import ProfilerActivity, profile, record_function
@@ -1351,6 +1352,33 @@ def parse_tool_response(text, tools, **kwargs):
         for call_info in call_info_list
     ]
     return text, call_info_list
+
+
+# TODO maybe move
+def weight_loader_tp_narrow(w: torch.Tensor, dim: int, start: int, length: int):
+    if isinstance(w, DTensor):
+        tp_device_mesh = get_tp_group().device_mesh_device
+        print(
+            f'weight_loader_narrow START {w.shape=} {w.dtype=} {type(w)=} {dim=} {start=} {length=} {w.device_mesh=} {w.placements=} {tp_device_mesh=} {w.device_mesh == tp_device_mesh=}')
+
+        ans = w
+        # TODO Remove this when one day the torch error "Cross device mesh comm not supported yet!" is implemented
+        ans = DTensor.from_local(ans.full_tensor(), device_mesh=tp_device_mesh,
+                                 placements=[Replicate() for _ in range(tp_device_mesh.ndim)])
+        ans = ans.redistribute(tp_device_mesh, [Shard(dim)]).to_local()
+
+        rank_via_mesh = tp_device_mesh.get_local_rank()
+        rank_via_arg = start // length
+        size_via_mesh = tp_device_mesh.size()
+        size_via_arg = w.shape[dim] // length
+        print(f'weight_loader_narrow END {rank_via_mesh=} {size_via_mesh=} {ans.shape=}')
+        assert rank_via_mesh == rank_via_arg
+        assert size_via_mesh == size_via_arg
+        assert ans.shape[dim] == length
+
+        return ans
+    else:
+        return w.narrow(dim, start, length)
 
 
 class MultiprocessingSerializer:
